backend/requirements.txt
asgiref==3.2.10
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==3.0.5
django-cors-headers==3.9.0
django-extra-fields==3.0.2
django-filter==2.4.0
django-templated-mail==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.2.0
exceptiongroup==1.1.1
gunicorn==20.0.4
idna==3.4
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.1.0
pytest==7.3.1
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2020.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.3.1
tomli==2.0.1
urllib3==2.0.3